requirements.txt
alembic==1.13.1
ansi2html==1.8.0
black==22.8.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
flake8==6.1.0
Flask==2.2.5
Flask-DebugToolbar==0.14.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.4
importlib-metadata==6.8.0
infinity==1.5
intervals==0.9.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Mako==1.3.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.6.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.4
packaging==21.3
pathspec==0.11.2
platformdirs==3.11.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.29
tenacity==8.2.3
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
validators==0.28.0
Werkzeug==2.2.3
WTForms==3.1.2
WTForms-Components==0.10.5
zipp==3.17.0
